Setup image policies
TL;DR
-----

Document set up for some simple image policies for staging and
production

Details
-------

Show the `tmc` commands to create image policies for the staging
and production workspaces. They are set up like the other TMC
objects in the demo with the native `tmc` templating capability.

For the staging registry, we show a set of trusted registries to
pull from. In this case, we show a trusted vendor (VMware, via
`registry.pivotal.io`), a more trustworthy public registry
(Google), and our private registry. We include the GCR registry
so that we can use the vulnerable "Kubernetes Up and Running" demo
image at the core of the demo.

For production, we only allow pulling from the private registry.