Fix broken link to basic setup (#11)
TL;DR
-----

Remove stray newline that broke the link to the setup instructions

Details
-------

Reviewing the docs with my team I noticed that the link to the
setup instructions was showing the Markdown text instead of a
link. Turns out it was a stray newline between the bracket
closing the link text and the paren opening the URI. Getting
rid of that newline fixed things up.