Document multi-cluster demo setup (#13)
TL;DR
-----

Document using multiple clusters

Details
-------

The original incarnation of this demonstration used multiple
clusters to better reflect the real world. When I decided to
document it for others to share, I started with a simplified
version that could work with a single cluster.

One of the first questions I got when presenting to a broader
team was why it didn't directly show that workspaces are able
to cross clusters. To address that concern, I've now laid out
instructions for the multiple cluster setup. I also did a bit
of refactoring of the documents to prevent repeating myself.

Lastly, validating the multi-cluster documentation surfaced
some errors in the single-cluster documentation that I've 
corrected with this change.